requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
nest-asyncio
pickleshare
PyJWT
python-decouple
pyOpenSSL
passlib